backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
redis==5.0.8
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.5.2
python-multipart==0.0.9
httpx==0.27.2
dnspython==2.6.1
python-whois==0.9.4
cryptography==43.0.1
scikit-learn==1.5.1
pandas==2.2.2
numpy==2.1.1
joblib==1.4.2
shap==0.46.0
